requirements.txt
langchain
python-dotenv
langchain_community
langchain_core
sentence_transformers
langchain-ollama
streamlit